+++ requirements.txt
@@ -1,5 +1,21 @@
-psycopg2-binary
-flask
+blinker==1.6.2
+click==8.1.3
+exceptiongroup==1.1.1
+Flask==2.3.2
+Flask-Cors==3.0.10
+Flask-SQLAlchemy==3.0.3
+greenlet==2.0.2
+iniconfig==2.0.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.3
+packaging==23.1
+pluggy==1.0.0
+psycopg2-binary==2.9.6
+pytest==7.3.2
+six==1.16.0
+SQLAlchemy==2.0.16
+tomli==2.0.1
+typing_extensions==4.6.3
 urllib3==1.26.6
-flask-cors
-pytest
+Werkzeug==2.3.6
