+++ rockr/__init__.py
@@ -1,33 +1,28 @@
 from flask import Flask
 from flask_cors import CORS
+from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy import URL
+from .settings import DATABASE_CONFIG
+
+db = SQLAlchemy()
 
 
 def create_app(test_config=None):
     # create and configure the app
     app = Flask(__name__, instance_relative_config=True)
     CORS(app)
-    # Don't need below quite yet
-    # app.config.from_mapping(
-    #     SECRET_KEY='dev',
-    #     # DATABASE=os.path.join(app.instance_path, 'flaskr.sqlite'),
-    # )
 
-    # if test_config is None:
-    #     # load the instance config, if it exists, when not testing
-    #     app.config.from_pyfile('config.py', silent=True)
-    # else:
-    #     # load the test config if passed in
-    #     app.config.from_mapping(test_config)
-
-    # # ensure the instance folder exists
-    # try:
-    #     os.makedirs(app.instance_path)
-    # except OSError:
-    #     pass
+    # CITE: https://flask-sqlalchemy.palletsprojects.com/en/3.0.x/quickstart/
+    # configure the database
+    url_object = URL.create(**DATABASE_CONFIG)
+    app.config["SQLALCHEMY_DATABASE_URI"] = url_object
+    # print(engine)
+    db.init_app(app)
 
     return app
 
 
 app = create_app()
 
+# import here to avoid circular imports. Not a great practice, but docs say it's okay
 import rockr.views
