+++ rockr/views.py
@@ -1,6 +1,6 @@
 from flask import request
 from rockr import auth, app
-from rockr.db.models import bands, users
+from rockr.models import User
 
 
 def format_response(status, data):
@@ -26,13 +26,13 @@ def logout():
 
 @app.route('/register', methods=['POST'])
 def register():
-    data = users.create_user(request.json)
+    data = User.create_user(request.json)
     return format_response(200, data)
 
 
 @app.route('/get_users', methods=["GET"])
 def get_users():
-    data = users.get_users()
+    data = User.get_users()
     return format_response(200, data)
 
 
@@ -44,5 +44,5 @@ def get_bands():
 
 @app.route('/get_user_role', methods=["POST"])
 def get_user_role():
-    data = users.get_user_role(request.json)
+    data = User.get_user_role(request.json)
     return format_response(200, data)
